safe remove

+++ superset/connectors/sqla/models.py
@@ -2043,7 +2043,8 @@ class SqlaTable(Model, BaseDatasource):  # pylint: disable=too-many-public-metho
             session.query(NewDataset).filter_by(sqlatable_id=target.id).one_or_none()
         )
         for tbl in dataset.tables:
-            session.delete(tbl)
+            if len(tbl.datasets) == 1 and tbl.datasets[0] == dataset:
+                session.delete(tbl)
 
         if dataset:
             session.delete(dataset)
